+++ packages/ht-angular/src/app/actions-status-graph/actions-status-graph.service.ts
@@ -4,7 +4,7 @@ import {DateRangeMap} from "ht-data";
 import { actionsClientFactory} from "ht-client";
 import { IActionStatusGraph} from "ht-models";
 import {filter, map} from "rxjs/operators";
-import {moment} from "moment-mini";
+import {format} from "date-fns";
 import {IActionsTrendlineConfig} from "../interfaces/trendline";
 import {ActionsGraph} from "ht-client";
 import {IAnalyticsItem, IAnalyticsItemService} from "../interfaces/analytics-item";
@@ -49,7 +49,8 @@ export class ActionsStatusGraphService implements IAnalyticsItemService {
   private getCompletedActionChart(data: IActionStatusGraph[]) {
     // const format = data.length < 15 ? 'MMM D' : "MMM D";
     const labels = data.map((item) => {
-      return moment(item.created_date).format('ddd, MMM Do')
+      return format(item.created_date, 'ddd, MMM Do')
+      // return moment(item.created_date).format('ddd, MMM Do')
     });
     const datasets = this.chartFormat.map((item) => {
       return {